requirements.txt
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
db-dtypes==1.4.3
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.34.0
google-cloud-bigquery-storage==2.32.0
google-cloud-core==2.4.3
google-cloud-language==2.17.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
idna==3.10
numpy==2.3.1
packaging==25.0
pandas==2.3.1
proto-plus==1.26.1
protobuf==6.31.1
py4j==0.10.9.7
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
